+++ src/os_credits/credits/billing.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
 
+from decimal import Decimal
 from typing import TypeVar
 
 from os_credits.credits.base_models import UsageMeasurement
@@ -11,7 +12,7 @@ from os_credits.log import internal_logger
 MT = TypeVar("MT", bound=UsageMeasurement)
 
 
-def calculate_credits(measurement1: MT, measurement2: MT) -> float:
+def calculate_credits(measurement1: MT, measurement2: MT) -> Decimal:
     """
     High-level function to calculate the credits based on the differences of the two
     usage measurements.
